Internal DBMS Version number adjustment from MMMM SSS PP (1570 003 05) to MM MM MM SSSS PPPP (15 07 00 0003 0005) This accommodates for 2 number Minor/Maintenance releases (Postgres and MySQL) And larger SupportPackages / PatchLevels (which can now hold a typical build number of 4 numbers)

+++ src/com/asetune/CounterControllerNoOp.java
@@ -29,10 +29,10 @@ extends CounterControllerAbstract
 	}
 
 	@Override
-//	public void initCounters(Connection conn, boolean hasGui, int srvVersion, boolean isClusterEnabled, int monTablesVersion) throws Exception
+//	public void initCounters(Connection conn, boolean hasGui, long srvVersion, boolean isClusterEnabled, long monTablesVersion) throws Exception
 //	{
 //	}
-	public void initCounters(DbxConnection conn, boolean hasGui, int srvVersion, boolean isClusterEnabled, int monTablesVersion) throws Exception
+	public void initCounters(DbxConnection conn, boolean hasGui, long srvVersion, boolean isClusterEnabled, long monTablesVersion) throws Exception
 	{
 	}
