+++ src/com/asetune/cm/rs/CmAdminStatsRsi.java
@@ -30,8 +30,8 @@ extends CmAdminStatsAbstract
 	public static final String   GROUP_NAME       = MainFrameRs.TCP_GROUP_MC;
 	public static final String   GUI_ICON_FILE    = "images/"+CM_NAME+".png";
 
-	public static final int      NEED_SRV_VERSION = 0;
-	public static final int      NEED_CE_VERSION  = 0;
+	public static final long     NEED_SRV_VERSION = 0;
+	public static final long     NEED_CE_VERSION  = 0;
 
 	public static final String[] MON_TABLES       = new String[] {"stats"};
 	public static final String[] NEED_ROLES       = new String[] {};
@@ -94,7 +94,7 @@ extends CmAdminStatsAbstract
 	}
 
 	@Override
-	public String[] getDependsOnConfigForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public String[] getDependsOnConfigForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		return NEED_CONFIG;
 	}
