+++ src/com/asetune/cm/rs/CmWsRepLatency.java
@@ -48,8 +48,8 @@ extends CountersModel
 	public static final String   GROUP_NAME       = MainFrame.TCP_GROUP_SERVER;
 	public static final String   GUI_ICON_FILE    = "images/"+CM_NAME+".png";
 
-	public static final int      NEED_SRV_VERSION = 0;
-	public static final int      NEED_CE_VERSION  = 0;
+	public static final long     NEED_SRV_VERSION = 0;
+	public static final long     NEED_CE_VERSION  = 0;
 
 	public static final String[] MON_TABLES       = new String[] {"logical_status"};
 	public static final String[] NEED_ROLES       = new String[] {};
@@ -215,7 +215,7 @@ extends CountersModel
 
 
 	@Override
-	public void addMonTableDictForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public void addMonTableDictForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		try 
 		{
@@ -279,7 +279,7 @@ extends CountersModel
 
 
 	@Override
-	public List<String> getPkForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public List<String> getPkForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 //		List <String> pkCols = new LinkedList<String>();
 //
@@ -299,7 +299,7 @@ extends CountersModel
 	}
 
 	@Override
-	public String getSqlForVersion(Connection conn, int srvVersion, boolean isClusterEnabled)
+	public String getSqlForVersion(Connection conn, long srvVersion, boolean isClusterEnabled)
 	{
 		// The SQL ins actually in: CounterSampleWsIterator
 		String sql =
